Added errors processing in handlers

+++ internal/handlers/joke.go
@@ -8,6 +8,8 @@ import (
 	"github.com/Sakagam1/DBMS_TASK/internal/db"
 	"github.com/Sakagam1/DBMS_TASK/internal/models"
 	"github.com/gorilla/mux"
+
+	customHTTP "github.com/Sakagam1/DBMS_TASK/internal/http"
 )
 
 func CreateJokeHandler(w http.ResponseWriter, r *http.Request) {
@@ -15,49 +17,85 @@ func CreateJokeHandler(w http.ResponseWriter, r *http.Request) {
 	var joke models.Joke
 	err := decoder.Decode(&joke)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	id, err := db.JokeRepo.Create(&joke)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
 	w.Header().Set("Content-type", "application/json")
+	w.WriteHeader(http.StatusCreated)
 	json.NewEncoder(w).Encode(id)
 }
 
 func DeleteJokeHandler(w http.ResponseWriter, r *http.Request) {
 	decoder := json.NewDecoder(r.Body)
 	var joke_id int
+	var user_id int
 	err := decoder.Decode(&joke_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
+	}
+	err = decoder.Decode(&user_id)
+	if err != nil {
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
+	}
+	user, err := db.UserRepo.GetUserByID(user_id)
+	if err != nil {
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
+	}
+	joke, err := db.JokeRepo.GetJokeByID(joke_id)
+	if err != nil {
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
+	}
+	if user.Role != "admin" && joke.AuthorId != user_id {
+		customHTTP.NewErrorResponse(w, http.StatusForbidden, "Error: "+err.Error())
+		return
 	}
 	err = db.JokeRepo.Delete(joke_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
-	w.Header().Set("Content-type", "application/json")
-	json.NewEncoder(w).Encode(err)
+	w.WriteHeader(http.StatusOK)
 }
 
 func GetUserJokesHandler(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	user_id, err := strconv.Atoi(params["userID"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	page, err := strconv.Atoi(params["page"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	pageSize, err := strconv.Atoi(params["pageSize"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	sortMode := params["sort"]
+	if sortMode != "hour" && sortMode != "day" && sortMode != "week" && sortMode != "month" && sortMode != "all" {
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
+	}
 	var jokes []models.Joke
 	jokes, err = db.JokeRepo.GetUserJokes(user_id, page, pageSize, sortMode)
+	if err != nil {
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
+	}
 	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(jokes)
 }
 
@@ -65,16 +103,27 @@ func GetPageOfJokesHandler(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	page, err := strconv.Atoi(params["page"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	pageSize, err := strconv.Atoi(params["pageSize"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	sortMode := params["sort"]
+	if sortMode != "hour" && sortMode != "day" && sortMode != "week" && sortMode != "month" && sortMode != "all" {
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
+	}
 	var jokes []models.Joke
 	jokes, err = db.JokeRepo.GetPageOfJokes(page, pageSize, sortMode)
+	if err != nil {
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
+	}
 	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(jokes)
 }
 
@@ -83,18 +132,26 @@ func SearchJokesByTagHandler(w http.ResponseWriter, r *http.Request) {
 	tag_name := params["tag"]
 	page, err := strconv.Atoi(params["page"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	pageSize, err := strconv.Atoi(params["pageSize"])
-	sortMode := params["sort"]
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
+	}
+	sortMode := params["sort"]
+	if sortMode != "hour" && sortMode != "day" && sortMode != "week" && sortMode != "month" && sortMode != "all" {
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	jokes, err := db.JokeRepo.GetJokesByTag(tag_name, page, pageSize, sortMode)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
 	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(jokes)
 }
 
@@ -103,18 +160,26 @@ func SearchJokesByKeywordHandler(w http.ResponseWriter, r *http.Request) {
 	keyword := params["keyword"]
 	page, err := strconv.Atoi(params["page"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	pageSize, err := strconv.Atoi(params["pageSize"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	sortMode := params["sort"]
+	if sortMode != "hour" && sortMode != "day" && sortMode != "week" && sortMode != "month" && sortMode != "all" {
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
+	}
 	jokes, err := db.JokeRepo.GetJokesByKeyword(keyword, page, pageSize, sortMode)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
 	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(jokes)
 }
 
@@ -122,60 +187,76 @@ func AddToFavoriteHandler(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	joke_id, err := strconv.Atoi(params["jokeID"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	decoder := json.NewDecoder(r.Body)
 	var user_id int
 	err = decoder.Decode(&user_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	err = db.JokeRepo.AddToFavorite(user_id, joke_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(err)
+	w.WriteHeader(http.StatusOK)
 }
 
 func DeleteFromFavoriteHandler(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	joke_id, err := strconv.Atoi(params["jokeID"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	decoder := json.NewDecoder(r.Body)
 	var user_id int
 	err = decoder.Decode(&user_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	err = db.JokeRepo.DeleteFromFavorite(user_id, joke_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(err)
+	w.WriteHeader(http.StatusOK)
 }
 
 func GetUserFavoriteJokesHandler(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	user_id, err := strconv.Atoi(params["ID"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	page, err := strconv.Atoi(params["page"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	pageSize, err := strconv.Atoi(params["pageSize"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	sortMode := params["sort"]
+	if sortMode != "hour" && sortMode != "day" && sortMode != "week" && sortMode != "month" && sortMode != "all" {
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
+	}
 	var jokes []models.Joke
 	jokes, err = db.JokeRepo.GetUserFavoriteJokes(user_id, page, pageSize, sortMode)
+	if err != nil {
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
+	}
 	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(jokes)
 }
 
@@ -185,18 +266,20 @@ func SubscribeToUserHandler(w http.ResponseWriter, r *http.Request) {
 	var sender_id int
 	err := decoder.Decode(&receiver_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	err = decoder.Decode(&sender_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	err = db.JokeRepo.SubscribeToUser(receiver_id, sender_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(err)
+	w.WriteHeader(http.StatusOK)
 }
 
 func UnSubscribeToUserHandler(w http.ResponseWriter, r *http.Request) {
@@ -205,18 +288,20 @@ func UnSubscribeToUserHandler(w http.ResponseWriter, r *http.Request) {
 	var sender_id int
 	err := decoder.Decode(&receiver_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	err = decoder.Decode(&sender_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	err = db.JokeRepo.UnSubscribeFromUser(receiver_id, sender_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(err)
+	w.WriteHeader(http.StatusOK)
 }
 
 func GetUserSubscribedJokesHandler(w http.ResponseWriter, r *http.Request) {
@@ -225,22 +310,31 @@ func GetUserSubscribedJokesHandler(w http.ResponseWriter, r *http.Request) {
 	var receiver_id int
 	err := decoder.Decode(&receiver_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	page, err := strconv.Atoi(params["page"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	pageSize, err := strconv.Atoi(params["pageSize"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	sortMode := params["sort"]
+	if sortMode != "hour" && sortMode != "day" && sortMode != "week" && sortMode != "month" && sortMode != "all" {
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
+	}
 	jokes, err := db.JokeRepo.GetUserSubribedJokes(receiver_id, page, pageSize, sortMode)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
 	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(jokes)
 }
 
@@ -248,52 +342,59 @@ func AddTagToJokeHandler(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	joke_id, err := strconv.Atoi(params["jokeID"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	decoder := json.NewDecoder(r.Body)
 	var tag_id int
 	err = decoder.Decode(&tag_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	err = db.JokeRepo.AddTagToJoke(joke_id, tag_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(err)
+	w.WriteHeader(http.StatusOK)
 }
 
 func DeleteTagToJokeHandler(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	joke_id, err := strconv.Atoi(params["jokeID"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	decoder := json.NewDecoder(r.Body)
 	var tag_id int
 	err = decoder.Decode(&tag_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	err = db.JokeRepo.DeleteTagFromJoke(joke_id, tag_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(err)
+	w.WriteHeader(http.StatusOK)
 }
 
 func GetJokeByIDHandler(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	joke_id, err := strconv.Atoi(params["jokeID"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	joke, err := db.JokeRepo.GetJokeByID(joke_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
 	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(joke)
 }
