+++ internal/handlers/report.go
@@ -9,6 +9,8 @@ import (
 
 	"github.com/Sakagam1/DBMS_TASK/internal/db"
 	"github.com/Sakagam1/DBMS_TASK/internal/models"
+
+	customHTTP "github.com/Sakagam1/DBMS_TASK/internal/http"
 )
 
 func CreateReportHandler(w http.ResponseWriter, r *http.Request) {
@@ -16,49 +18,79 @@ func CreateReportHandler(w http.ResponseWriter, r *http.Request) {
 	var report models.Report
 	err := decoder.Decode(&report)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	id, err := db.ReportRepo.Create(&report)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
 	w.Header().Set("Content-type", "application/json")
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(id)
 }
 
 func DeleteReportHandler(w http.ResponseWriter, r *http.Request) {
-	params := mux.Vars(r)
-	report_id, err := strconv.Atoi(params["reportID"])
+	decoder := json.NewDecoder(r.Body)
+	var report_id int
+	var user_id int
+	err := decoder.Decode(&report_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
+	}
+	err = decoder.Decode(&user_id)
+	if err != nil {
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
+	}
+	report, err := db.ReportRepo.GetReportByID(user_id)
+	if err != nil {
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
+	}
+	user, err := db.UserRepo.GetUserByID(user_id)
+	if err != nil {
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
+	}
+	if user.Role != "admin" && user_id != report.SenderId {
+		customHTTP.NewErrorResponse(w, http.StatusForbidden, "Error: "+err.Error())
+		return
 	}
 	err = db.ReportRepo.Delete(report_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
-	w.Header().Set("Content-type", "application/json")
-	json.NewEncoder(w).Encode(err)
+	w.WriteHeader(http.StatusOK)
 }
 
 func GetReportByIDHandler(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	report_id, err := strconv.Atoi(params["reportID"])
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusBadRequest, "Error: "+err.Error())
+		return
 	}
 	reportOut, err := db.ReportRepo.GetReportByID(report_id)
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
 	w.Header().Set("Content-type", "application/json")
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(reportOut)
 }
 
 func GetAllReportsHandler(w http.ResponseWriter, r *http.Request) {
 	reportOut, err := db.ReportRepo.GetAllReports()
 	if err != nil {
-		panic(err)
+		customHTTP.NewErrorResponse(w, http.StatusInternalServerError, "Error: "+err.Error())
+		return
 	}
 	w.Header().Set("Content-type", "application/json")
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(reportOut)
 }
